Added float to circle area.

+++ ch3/circle_area.java
@@ -2,9 +2,9 @@ class	circle_area	{
 	public static void main(String args[])	{
 
 		int radius = 10;
-		double pi, area;
+		float pi, area;
 
-		pi = 3.1416;
+		pi = 3.1416f;
 		area = pi * radius * radius;
 
 		System.out.println("Area : " + area);
